Use own component for Yeti layer
